revert find_delegation_validators -> no state
